Include import subscriptions popup in the user_profile layout.
This layout is used in the edit registration view. This view is accessible while the user is authenticated; therefore the user dropdown with its "import subscriptions" link is visible. This means the popup markup must be included in the page, otherwise the link does not work.

Not that I expect many users to access it from this view, but if the option is present in the menu, it must be functional.