Ensure that when adding the first folder or feed to the root scope in angularjs, any ng-repeat directives are triggered.
To do this, if the added element (feed or folder) is the first in the array, the array is recreated as a new array with only that element. For some reason it seems that just pushing an element into an already existing empty array does not trigger ng-repeat directives. Maybe an angularjs bug?

This fixes the issue in which if a user who had no subscriptions added a subscription, it did not appear in the sidebar until the next page refresh.