Displaying certain view elements (unread count badges, entries list under folders...) is now based on the value of a boolean flag in the root scope.
The old way of doing it (checking for the presence of the feeds list in the root scope) failed in the case in which the server returned an empty feeds list. This is very relevant because this affected all new users, for whom some parts of the application did not display properly until the first page reload (e.g. the unread entries count in the "all subscriptions" folder wasn't displayed). This is very bad, a bad first impression could scare away those users for good.