Instead of an "all subscriptions" folder, a stacked-pill list is displayed in the sidebar above any folders, with those feeds that are not in any folder.
Also an "all subscriptions" link is displayed in that list which, when clicked, loads the list of entries from all subscribed feeds.

This should make the client be a bit faster. The way it was done before, a lot of feeds were duplicated in the sidebar (once in the "all subscriptions" folder and then again in their own folders). In the extreme case all feeds could be duplicated. This means added overhead from angularjs watches etc.

Of course this breaks a lot of acceptance tests. Next in line: fixing them. Who said refactoring is fun? Not me!