Folders without unread entries are hidden by default. Their visibility is toggled with the "show read feeds" navbar button.
All folders are actually present in the JSON returned by FoldersController#index; this is necessary so that all folders appear in the folder dropdown, even those without unread entries (we want the user to be able to move a feed to any folder, including those which currently have no unread entries). Angularjs takes care of triggering the visibility of folders based on the "show_read" global flag and the presence or absence of unread entries in each folder.