Do not hide the folder of the currently selected feed, or the currently selected folder, even if the folder has no unread entries.
This keeps the whole folder from disappearing if the user marks as read all entries in the single feed inside a folder.

Also, do not check for folder removal when cleaning read feeds from the sidebar. Folders should never be removed from the model in this case, as they actually still exist and should appear in the folder dropdown. The angularjs filter takes care of not displaying folders without entries in the sidebar, but they are not actually removed from the model.