Removed two unnecessary lines that should have been removed. Angularjs services that retrieve the currently selected feed or folder always return an object present in the scope.
Angularjs services are now aware that the feed and folders list can have changed after the current feed or folder was selected and stored in the root scope (e.g. loading read feeds or removing a feed from a folder trigger AJAX reloads of the feeds and folders list, respectively). If the object stored in the root scope as current feed or folder was naively returned, it could happen that the object was not actually in the corresponding list; if the list has been reloaded, there will be another object with the same ID, but a different object nonetheless.

This returning of a different object breaks meny things in the view.

The fix is, every time that the current feed or folder is requested, search in the feeds or folders list for an object with the same ID as the one stored in the root scope, and return that object, thereby making sure that the returned object is in the list.