Reorganization of the angularjs code related to showing/hiding read feeds.
Now instead of passing a boolean argument around, a boolean flag is set as soon as possible to indicate whether we want to load read feeds or not, and most methods just read the flag from the root scope instead of expecting a boolean argument.

This is the first step towards extending the functionality of this flag, so that it also controls whether to load read entries.