Users can use the "Read All" button to mark all entries as read while reading a whole folder, or even the "all subscriptions" list.
The application is smart enough to change only those entries the user is seeing (and older ones). If after the user loaded the list of entries a feed refresh happened in the background which retrieved new (unread by default) entries, those new entries (which are not currently visible to the user) will not be affected when the user clicks on Read All.