Opening and closing entries is not achieved with Bootstrap animations anymore. Angularjs conditional classes are used instead.
This means that the smooth open/close animation is lost, and the transition between open/closed states is instantaneous. Maybe in the future we can try to find a way to get animations back, but for the moment implementing the "scroll entry to make it as visible as possible" functionality is more important.