When opening an entry, automatically scroll so that the entry header is at the top of the viewport, with another entry header visible above it.
This helps usability: when the user opens an entry, the app tries to scroll to show as much of its body as possible inside the viewport. This removes the need to scroll every time a user opened an entry to start reading it from the top.

This has been performed with the $location and $anchorScroll angularjs services. This means that the hash fragment of the current URL is changed every time the user opens an entry. It would be nice not to have this behavior, but I can see no easy way of hiding this from the user.