When fetching a feed, already existing entries are updated but their publish date is not changed.
This is intended for the case in which a feed returns no publish date for its entries (therefore the fetch date is used by default) and also it doesn't support HTTP caching (it returns the full entries list every time it's fetched). If we updated the publish date in this case, the publish date of entries in this feed would be updated to the current datetime every time the feed was fetched, which would mean entries from this feed would always "float" to the top of the entries list. The fix is to use the publish date the entry had the first time it was fetched, and every time it's fetched again keep the original publish date.