Every time a scheduled feed update runs, the unread entries count is corrected if necessary.
This slows down updates a bit, but we ensure that the unread entries count in each subscription is periodically checked and corrected if necessary. This should reduce the cases in which the app reports a different count of unread entries than the number of unread entries actually retrieved when reading the feed.