Fixed a bug in the angularjs service that finds feeds in a folder, which sometimes caused the wrong feeds to be returned.
The problem is that the $filter('filter') builtin angularjs method was being used incorrectly. The way it was used, it performed a subscring match between the id of the folder and the folder_id attribute of each feed; this meant that e.g. when searching for feeds in folder with id "4", feeds in folder with id "14" will be returned as well.

The solution is to use the builtin filter method correctly. Instead of passing an options object for the third argument, which performs a substring match, a function is passed which compares the folder.id and feed.folder_id attributes and only returns true if they are equal.

Note that the method compares ids, not objects themselves, and therefore works fine even after an AJAX reload of the feeds or folders lists, which discards all objects and inserts new ones in the lists.