While the page is open, feeds are reloaded every minute.
Updated unread counts appear in the sidebar and any feeds previously didn't have any unread entries but now do (because a scheduled refresh has been run in the background) appear automatically in the sidebar, along with their folders if they weren't already visible.

We must ensure that if feeds have not changed, a 304 is returned instead of the same JSON list, to improve performance.

As a side effect of changes in this commit, when an OPML import finishes the "show_unread" is not reset to false. This could have surised affected users. Also when such an import finishes new feeds and folders are inserted in the sidebar without showing an animated spinner in it.