When retrieving feeds via AJAX, feeds already in the scope are updated, new feeds are inserted in scope, but if a feed is in scope but it's not in the returned JSON, it's left in the scope untouched.
This fixes the following situation: with the old behaviour (the retrieved feeds list simply replaces the feeds list in scope), if a user marked as read all entries in a feed and, keeping that feed selected, waited for the next automatic reload of entries (every minute), the feed he had selected disappeared from the sidebar.

There are already mechanisms in place for removing feeds from the sidebar when they have no unread entries and the user selects a different feed, as long as he's not toggled the "show read" button. Feeds should not disappear from the sidebar because of the automatic refresh every minute (but they should appear in the sidebar automatically, if there's been a background fetch of a feed which had no unread entries and now it does have unread entries).