Fixed unsubscribe functionality.
With the recent changes, findSvc returns a null if it cannot find what it's been asked for (feeds in a folder in this case). This null return value was not properly supported in the angularjs code that is executed when the user clicks on "Unsubscribe".

I'll have to check if there are other places where this happens.