Added a method to SubscriptionsManager to recalculate the unread entries count of a feed for a user.
This new method actually counts database records, ensuring the unread entries count is correct.

It does not write to the database unless necessary (i.e. if the currently saved count is different from the calculated count).

Changed UpdateFeedJob so that it invokes this method, instead of having the code for this operation in the job.