WIP: perform cached unread entries count calculations with mass count operations when possible, instead of using activerecord callbacks every time an EntryState instance is created, destroyed, or changes state.
Working with activerecord callbacks to keep the unread entries count up to date is failure-prone and very slow. E.g. when subscribing to a feed, we were doing a SQL update for EACH entry in the feed, to increment the unread_entries field by 1 every time. It is much more efficient to remove those callbacks and, when subscribing, just do a SQL count and a single SQL update after adding the feed to the user's subscpritions (thanks to callbacks, when adding a feed to a user's subscriptions, an EntryState with read=false is created for each entry in the feed).