In angularjs code, when setting the unread count for a feed to zero, first search for it in the feeds list stored in the root scope.
Previously the angularjs service acted on the currently selected feed. The problem is that if an automatic client refresh (every minute) has happened between selecting the feed and marking all its entries as read, the object for the feed could be a different object instance. We have to search the actual object for the list using its id, and then set the count to zero in that object.