WIP: fix acceptance tests.
A lot of test setups have been broken because they operated under the assumption that just by adding entries to a feed, the unread entries count for all subscribed users would be updated.

In most cases the fix is to subscribe users AFTER adding entries to feeds, during test setup.