Before removing the current feed from the angularjs scope during unsubscribing, retrieve the actual feed object from the current feeds list.
This is intended for the case in which between the user clicking on the feed and him clicking on "unsubscribe", an automatic refresh of the client (every minute) happens. This would mean that a feed object with the same id as the selected one will be in the scope, but it will not be the same object that was stored in the current_feed variable.