As much as reasonable, angularjs code accesses the currently selected feed through the currentFeedSvc service instead of directly accessing the $rootScope.current_feed variable.
This is because after an automated client refresh (every minute), the feeds list is recreated and the object stored in the $rootScope.current_feed variable is no longer in the feeds list. Going through the currentFeedSvc.get() function ensures that a feed object actually present in the feeds list is correctly returned.