When fetching a feed, entries already in the database will be ignored.
Previously the entries already in the db were updated with the data fetched from the feed. In the vast majority of cases these will be feeds that do not honor HTTP caching headers and send the full list of entries every time, and the information passed to the db update will be exactly the same already in the db, making such updates generally a waste of resources. It's reasonable simply to ignore them.