Reduced feed and folder title's maximum width in the sidebar.
In chrome the scrollbar for the sidebar is wider than in the Firefox I normally use. I realized that if the scrollbar is visible and a folder has "999+" in the unread entries badge, the badge "jumps down" to the next line.

Reducing the folder title width a bit avoids this. I haven't seen the same happening with folders, but I guess it could happen so the change has been aplied to them too.