Moved data initialization to as soon as possible during an acceptance test.
I'm getting an elusive failure on folders_spec.rb acceptance test. I think it's a timing problem: sometimes the test runs so fast that it opens the page before the data is actually written to the database; the result is that some folders that should be visible don't appear in the view.

By moving database initialization to as soon as possible during test setup I hope the test will pass more often. Hopefully always!