Reduced the number of alerts displayed to the user.
When a running AJAX request is cancelled because the user cancels it (e.g. by refreshing the page), no alert is displayed. This is not an error condition. Unfortunately this also means that if the request is cancelled because the server is down, no alert is displayed. However this should be obvious to the user in the most important cases (feeds not being loaded, entries not being loaded etc).

Also when an error happens while retrieving the import process status for a user, no error is displayed. This is a background process of which the user will be notified by email, and the AJAX progress-bar is not a critical part of the functionality. An error here is not grave enough to justify showing an alert.