Mass-marking several entries as read (a whole feed, a whole folder or all entries) is done with a single SQL UPDATE.
This replaces the old behavior of entering a loop of instancing an EntryState object, marking it as read and saving it back into the database for each affected entry. It should be much faster now.

ActiveRecord callbacks are not executed when mass-updating records with the update_all method, but this is not a concern because the EntryState model has no callbacks.