When user unsubscribes from a feed, immediately reload folders from server.
The client doesn't have the information to know if a folder has to be removed from the scope. Maybe the client knows of no feeds in the folder (besides the one just unsubscribed), but there could be other feeds without unread entries which the client currently is not aware of, and therefore the folder should not be removed in the client (it should still appear in the dropdown).

The only way to ensure data consistency is to ask the server which folders currently exist after unsubscribing from the feed.