Fixed acceptance tests.
Update of unread entries count in feeds, after marking multiple entries as read, now happens in a background job. This means that if the acceptance tests immediately refreshes the page after clicking on "mark all as read" button, the feeds affected will still be in the sidebar, because there is no running Resque during acceptance testing.

Removed all instances of refreshing the page to see if the feeds disappear after clicking on "mark all as read", in acceptance testing.