WIP: mark entries as read as soon as they leave the effective viewport.
The app has a fixed navbar which covers the top of the actual viewport. We want to mark entries as read as soon as they "slide" under the navbar, at which point they are still in the actual viewport but above the effective viewport (which extends down from the bottom of the navbar).

Jquery-viewport.js has been modified inspired by this commit:

	https://github.com/theluk/jquery_viewport/commit/d7ed501b5eac2bed90253c134df611cd2d843112

(the only difference is I have applied the changes directly on the minified library instead of the source code).

With this change, the viewport selectors added by jquery-viewport.js accept an offset parameter. With this offset we can detect entries that have slid under the navbar.