WIP: Quick Reading mode. Replaced jquery-viewport library with within-viewport library.
The viewport-selectors Jquery plugin:

	http://www.appelsiini.net/projects/viewport

did what I wanted, but it performs horribly. When scrolling, the page froze for a few instants.

Instead now I use within-viewport JS library (and its JQuery plugin):

	https://github.com/patik/within-viewport

It makes the code a bit more complex (because I'm not interested in entries within the viewport but rather outside it, which forces me to select all entries and then do a jquery "not" with the entries returned by within-viewport), but it has noticeably better performance. The library author seems to think about performance, because the library can be told to watch only one viewport edge (top, in my case) for better performance. Also it allows setting a top-offset to the viewport (in my case to account for the navbar) which fits my needs nicely.