Added acceptance tests for the Quick Reading mode feature.
These are not very exhaustive, just smoke tests. Also testing functionality related to scrolling with capybara is a pain in the ass; I get the feeling these tests will prove to be brittle. For now it's enough to consider the feature implemented and tested, but I'll keep an eye on improving tests.