Reverted commits 6b68047ab5942399ba6b5be9d85da374527f3333 and a7bdc1180120b0b3602a749adc3859d4af374f7d.
These commits are "Added not-null constraints and default values for many database columns" and "Removed default values from the definition of TEXT columns in the database" respectively.

It is necessary to revert these commits so that migration 20131120115233_add_database_constraints.rb disappears from the db/migrate directory.

The problem is that when deploying via capistrano in staging or production, pending database migrations are run sequentially and this particular migration fails (because it tries to set default values in TEXT columns, which is not supported by MySQL), making the whole deployment fail. This error breaks the deployment even though the next migration that would be run (20131121122420_remove_defaults_from_text_fields.rb) would remove those defaults. This does not work; it is not good enough to remove defaults in a later migration, because they cannot be set in the first place.

The solution is to remove both migrations reverting their commits. In a later commit a new migration will be added which will be similar to 20131120115233_add_database_constraints.rb but without setting defaults for TEXT fields.