Reverted commits d712f7b2d955f15ca7ca52612c27951d0eb15674 and d9208854d82f282552f60574e6a983f2b4542ccf
The change in the capistrano script to use "with" and "within" directives was not working properly. Specifically capistrano was not issuing a "cd" command into the current folder before trying to run a "bundle exec god terminate"; as a result, bundler could not find the Gemfile (because the command was not run from the correct folder) and the whole deployment failed.

The way is written now is a bit too verbose but works. It's not worth it to keep looking into this.