In God configuration, put resque-related watches (resque-worker, resque-scheduler) and the redis-server watch in different groups.
This is a first step towards only restarting resque watches when deploying the app, and leaving redis-server running all the time.

If a configuration change requires a restart of the redis server it can be done manually, of course. But it should not happen often. Resque workers however have to be restarted on each deploy, so that any changes in code are picked up (resque reads ruby code when started and does not read it again unless restarted).