Fixed God-related capistrano tasks.
Instead of manually running a "cd" command, "within" blocks are used.

Instead of manually setting the RESQUE_ENV environment variable when running commands, "with" blocks are used (and RAILS_ENV is added to the default environment in config/deploy/production.rb and staging.rb files, so it is used in all commands).

Instead of manually running "bundle exec god" every time, a :god key is added to the SSHKit command map so that the :god command can be executed by capistrano and actually a "bundle exec god" is executed in the remote servers.

Most of these are features of SSHKit, which is the library that powers Capistrano 3:

	https://github.com/capistrano/sshkit

An important limitation of SSHKit is that "execute" cannot be passed a command string with whitespaces in it. Instead several strings (without whitespaces) can be passed, each one with part of the whole command, and when running it in the remote servers SSHKit concatenates all of them with witespaces between them. It makes for an awkward syntax but devs seem to think it unavoidable. For more details see:

	https://github.com/capistrano/capistrano/issues/719