Every time a scheduled feed update is run, the current schedule is reviewed and any missing scheduled updated is added.
This is a safeguard against recent instances of feed update schedules simply disappearing. I'm not sure if the problem is with redis, with the overcommit configuration of the boxes, with the way redis and resque were being restarted or from some other source. All of the above have been reviewed and improved, and with this change schedules should be fixed as long as there is a single feed update scheduled. The price to pay is an overhead when running each scheduled update, as the schedules are reviewed and fixed if necessary.