Fixed resque (and resque-scheduler) configuration.
The rails initializer for resque only tries to load the static scheduler if running in a background server. This is intended to fix a bug in the CI environment: when loading the rails environment for running tests in the CI environment, in which there is no redis instance running, the initializer tried to access the configured redis instance to save the static schedule into it. This made the build fail in Travis CI because of an error trying to access a non-existing redis instance. This fix will keep Travis CI from trying to read the static schedule and save it into redis.

Also, the rake task that runs resque and resque-schedule workers has been refined. Now it shares all necessary configuration with the rails resque initializer (only simpler because the rake task always runs in a background server, so it is not necessary to check if the code is running in a web or a background server). Also all job classes are explicitly required, and the static schedule is loaded in redis if the rake task is run standalone, without an initialized rails environment (which is what happens in background servers normally).