Streamlined god configuration.
I'm seeing some inconsistent behavior from God in the staging and production environments. Specifically when examining running processes, I see that e.g. the redis server is using a configuration file path which points to the last release path at the time the process was started. This means that even after several redeploys, god is still using paths pointing to what was the latest release path when god was started; this is a problem because after several redeploys that folder could have been deleted (capistrano only keeps the last 5 releases by default).

I think the problem may be that even if god is told to reload its config every time capistrano redeploys the app, maybe environment variables are not passed to the new config file, but rather the values those env variables had when god was started are passed to the config file. The config file was written in such a way that it used env variables set by capistrano to find out the absolute path for the current release; if the old values of those env vars are used, that meant that even after loading a new config file God still made reference to old paths.

This is conjeture on my part, but that's what I think is happening. To fix this the god config file has been changed so that only env vars which do not change from deployment to deployment are used (specifically RESQUE_ENV, which is "background" for background servers, and RAILS_ENV). Paths are now hardcoded in the config file, depending on the rails environment, and should point to the correct folder in the "current" or "shared" capistrano folders.

As a side effect, several env variables are no longer passed from capistrano to god.