Fixed bug when fetching itunes feeds.
It was a bug in feedzirra gem; it did not support itunes feeds with a content.

Fixed by this pull request in feedzirra:

https://github.com/pauldix/feedzirra/pull/198

Currently not in any stable tag, temporarily used feedzirra's github master as a bundler dependency. When there is a stable version that includes this pull request, I'll use it in the Gemfile instead of the github master.

Not yet implemented: unit test to ensure that Feedbunch correctly fetches itunes feeds like the one above.