Reverted commit 343f45427a646c05bd1a7da05f8252c9b0bd9cdd and fix it in feedbunch, instead of relying on unreleased feedzirra gem versions.
Now the Gemfile specifies the last stable release of the feedzirra gem as a dependency. This version of feedzirra does not add a "content" attribute to entries in itunes feeds, which caused a problem in feedbunch when saving entries in the database.

The fix to make feedbunch work with stable feedzirra is to check if the entry object returned by feedzirra supports the "content" method. If it does, the entry is saved as before. If it doesn't, the entry content in feedbunch defaults to the entry's summary.

This should fix any potential problems with other feedzirra parsers which do not support the "content" attribute in entries (I don't know if there are any, but they will not be a problem in any case).