Restart God, Resque and Resque-scheduler every time the app is deployed with capistrano, but without restarting Redis.
This should fix errors after several capistrano redeployments without restarting God, because the god process was running in a bundler environment whose Gemfile no longer existed. Now God is restarted, which ensures it works fine with the latest deployment, but Redis is not restarted by default, which minimizes downtime and possibility of failures.