JSON-escape the GUID and URL of each entry before returning it to the client.
This should fix the problem with entries with characters that looked like invalidly JSON-escaped characters; e.g. an entry with this guid and url:

   "http://ehtio.es/comic_es/non_sequitur/1027/i\m-on-fire"

would be returned as JSON like:

{
    "id": "59",
    "title": "I\\'m on fire",
    "read": false,
    "published": "05 Jul 00:00",
    "url": "http://ehtio.es/comic_es/non_sequitur/1027/i\m-on-fire",
    "guid": "http://ehtio.es/comic_es/non_sequitur/1027/i\m-on-fire",
    "summary": "whatever",
    "feed_id": "2"
}

which is not actually valid JSON because of the "\m" character in the guid and url (it can be checked with jsonlint.com).

With the fix in this commit the above entry would be returned as:

{
  "id": "59",
  "title": "I\\'m on fire",
  "read": false,
  "published": "05 Jul 00:00",
  "url": "http://ehtio.es/comic_es/non_sequitur/1027/i\\m-on-fire",
  "guid": "http://ehtio.es/comic_es/non_sequitur/1027/i\\m-on-fire",
  "summary": "whatever",
  "feed_id": "2"
}

which is valid JSON.